+++ src/PortableDeveloper.Application/Php/IPhpSettingsStore.cs
@@ -1,0 +1,8 @@
+namespace PortableDeveloper.Application.Php;
+
+public interface IPhpSettingsStore
+{
+    PhpSettings Load(string instanceId = "default");
+
+    void Save(PhpSettings settings, string instanceId = "default");
+}